rust/hw/gpio: Add the the first gpio device pcf8574
After implementing the I2CBus and I2CSlave, we proceeded to implement a basic
GPIO device — the PCF8574 — which depends on I2CSlave.

If we had directly referred to the PL011 implementation, we would have been
able to achieve a basic version of the PCF8574. However, during the
implementation, the excessive number of types and conversions led to various
comprehension conflicts and cognitive overhead.

The most significant insight I gained during this process is that the current
approach in Rust appears to be more complex and harder to understand than its
C counterpart. This trend poses a substantial barrier for developers interested
in Rust for QEMU, preventing them from efficiently, quickly, and safely modeling
QEMU devices in Rust. In my opinion, this deviates from the original intent of
adopting Rust.

Moreover, the mechanisms of bindings and wrappers are overly redundant, and
there is no clear indication that they are mandatory. These constructs generate
a large number of similar structures in the final compilation artifacts. Could
we design a more generic approach to bindings and wrappers?

Additionally, vmstate has not been fully implemented for the PCF8574 (to be
precise, it is not yet complete). During the implementation, I encountered an
issue: initially, I tried to directly transliterate the C struct — that is, I
attempted to configure vmstate and handle other aspects directly via field
mappings. However, I eventually found that some internal APIs were incompatible.
To make it work, I had to isolate those "potentially mutable values" into a
separate structure, akin to something like a "PL011Register", in order for it
to function correctly. Moreover, the vmstate configuration also needed to be
declared separately. Although I could infer its meaning, I still hope that a
more rational and streamlined process can be established.

At present, I have summarized the general workflow for device modeling in Rust
as follows:

1. Modify the configuration under hw/deviceto distinguish between C and Rust
   versions.
2. Create a device crate under rust/hw.
3. Add (or copy) the necessary wrappers, build.rs, and bindings.
4. Begin the device modeling process.
5. Construct the corresponding structures in Rust that mirror those in C,
   especially for “members that may change”.
6. Referencing the C implementation, define initialization functions and
   establish parent-class relationships for the Rust structure.
7. Set up ObjectImpl, DeviceImpl, and ResettablePhasesImpl.
8. Configure vmstate.
9. Implement other functional methods.

Signed-off-by: chenmiao <chenmiao@openatom.club>